QMutex::unlock: replace testAndSet with fetchAndStore for futexes
For systems where we know for sure that we're using futexes, this moves
the storing of the nullptr (i.e., the unlocking) from unlockInternal()
up to unlock() itself. Thus, it replaces a compare-and-exchange
operation plus a store operation with a single one: a simpler exchange.

I don't expect measurable gains with this, as the branch into
unlockInternal() still depends on the result of an atomic operation.

Change-Id: I120dbdc21e678f4b858ffffd60e6ee013ffff555
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>