QDBusPendingReplyBase: add missing copy operations
QDBusPendingReplyBase declares its destructor (and defines it,
as defaulted, out-of-line). I'm not sure why that's the case, since
QDBusPendingReplyBase isn't polymorphic. This generates warnings when
QDBusPendingReplyBase objects are copied in the subclasses, since
the copy operations are deprecated due to the presence of the
destructor.
Add the missing copy operations, and mark destructor and copies
to be removed in Qt 7.

Change-Id: Iec2c98f5e512c3f19e5d28ab2ce2dba365c66175
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>