Fix bug QGridLayoutEngine::cellRect()
lastColumn and lastRow could exceed the total column/row count, leading
to out-of-bounds accesses to the QList

QGridLayoutEngine::cellRect() is unused, so there is no risk of
regressions here. We fix it because the function might be used later.

So far the plan is to use it for visualizing the layout cells in QtQuick
Layouts.

Change-Id: Idd9db682039f2f81b3a6fbcd03990b6f15334d3e
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>