tst_QApplication: test quit() not ::exit() from an event loop
Amends Qt 4.8 commit 8dd8db250d92521fda619bdcf3e1c859b37b2da0.

The report in QTBUG-12673 says "quit()" and the backtrace clearly shows
the issue was attempting to cleanly exit the event loop while a modal
dialog was showing. The test instead tested ::exit().

That was actually an important test because nothing else in the Qt test
suite was testing this, but now it's included in the exitFromEventLoop()
test for QCoreApplication, QGuiApplication, and QApplication.

Task-number: QTBUG-12673
Change-Id: Ie7bd19af7103f3066b10fffda299d1168c8f677e
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>