tst_QAccessibility: Use QKeySequence::toString() properly
The test gets a string representation of a key sequence by stitching
the individual parts manually. It is not a portable approach and the
test relies on undocumented behavior of QKeySequence::toString().

A more reliable and portable approach is to pass the Qt::Key value to
QKeySequence and let it do its magic.

Change-Id: I2a0e8734c93fa4f86f11324f26d8967d22e14247
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>