Accommodate more space for new shapes in Qt::CursorShape in Qt 7
Currently, the Qt::CursorShape enum can accommodate only two new shapes
because Qt::LastCursor must be less than Qt::BitmapCursor.

The cursor shapes change rarely but it would be nice to keep them open
for extension in order to be able to react to new UI/UX ideas if needed.

Change-Id: I53ed9d742a09bb06bab73a906643c62a77a207b0
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>